requirements.txt
absl-py==1.3.0
astunparse==1.6.3
autopep8==2.0.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
docutils==0.19
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
keras==2.11.0
keyring==23.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplot==0.1.9
matplotlib==3.6.2
more-itertools==9.0.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
pkginfo==1.8.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
readme-renderer==37.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
rsa==4.9
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
tomli==2.0.1
twine==4.0.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
